requirements.txt
Flask==3.0.0
openai==1.6.1
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.31.0
